+++ .vuepress/enhanceApp.js
@@ -35,8 +35,10 @@ const iconsToMock = [
   'faLightbulb',
   'faThumbsUp',
   'faEllipsisV',
-  'faTimes'
+  'faTimes',
+  'faCheck'
 ]
+
 const mockedFalIcons = _.mapValues(_.pick(fas, iconsToMock), function (original) {
   return _.assign({}, original, {
     prefix: 'fal'