+++ test/unit/specs/elements/GeoEditableInput.spec.js
@@ -1,12 +1,28 @@
+import _ from 'lodash'
 import { createLocalVue, mount } from '@vue/test-utils'
 import GeoActivityIndicator from '@/elements/GeoActivityIndicator/GeoActivityIndicator.vue'
 import GeoDropdown from '@/elements/GeoDropdown/GeoDropdown.vue'
 import GeoEditableInput from '@/elements/GeoEditableInput/GeoEditableInput.vue'
+import GeoCompactButton from 'src/elements/GeoCompactButton/GeoCompactButton.vue'
+import GeoPrimaryCompactButton from 'src/elements/GeoCompactButton/GeoPrimaryCompactButton.vue'
+import GeoSecondaryCompactButton from 'src/elements/GeoCompactButton/GeoSecondaryCompactButton.vue'
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import { library } from '@fortawesome/fontawesome-svg-core'
 import { fas } from '@fortawesome/free-solid-svg-icons'
+import { FontAwesomeIconMock, expectFontAwesomeIconProp } from 'test/unit/utils/FontAwesomeIconMock'
 
-library.add(fas)
+const iconsToMock = [
+  'faTimes',
+  'faCheck',
+  'faBell'
+]
+const mockedFalIcons = _.mapValues(_.pick(fas, iconsToMock), (original) => {
+  return _.assign({}, original, {
+    prefix: 'fal'
+  })
+})
+library.add(mockedFalIcons)
+
 // create an extended `Vue` constructor
 const localVue = createLocalVue()
 localVue.component('geo-activity-indicator', GeoActivityIndicator)
@@ -17,12 +33,9 @@ describe('GeoEditableInput', () => {
   it('Should render input component', function () {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: false,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: false
       },
       stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
         GeoDropdown
       }
     })
@@ -33,12 +46,9 @@ describe('GeoEditableInput', () => {
   it('Should emit an event on click', function (done) {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: false,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: false
       },
       stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
         GeoDropdown
       }
     })
@@ -54,16 +64,49 @@ describe('GeoEditableInput', () => {
     })
   })
 
+  it('Should render correct icon for saveIcon when provided', function () {
+    const wrapper = mount(GeoEditableInput, {
+      propsData: {
+        showButtons: true,
+        saveIcon: ['fas', 'thumbs-up']
+      },
+      stubs: {
+        GeoDropdown,
+        GeoCompactButton,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton,
+        'font-awesome-icon': FontAwesomeIconMock
+      }
+    })
+    const fontAwesomeIconElem = wrapper.find('.geo-compact-button--primary').find(FontAwesomeIconMock)
+    expectFontAwesomeIconProp(fontAwesomeIconElem, ['fas', 'thumbs-up'])
+  })
+
+  it('Should render correct icon for cancelIcon when provided', function () {
+    const wrapper = mount(GeoEditableInput, {
+      propsData: {
+        showButtons: true,
+        cancelIcon: ['fas', 'thumbs-up']
+      },
+      stubs: {
+        GeoDropdown,
+        GeoCompactButton,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton,
+        'font-awesome-icon': FontAwesomeIconMock
+      }
+    })
+    const fontAwesomeIconElem = wrapper.find('.geo-compact-button--secondary').find(FontAwesomeIconMock)
+    expectFontAwesomeIconProp(fontAwesomeIconElem, ['fas', 'thumbs-up'])
+  })
+
   it('Should not emit the event on click when disabled', function (done) {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
         showButtons: false,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check'],
         disabled: true
       },
       stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
         GeoDropdown
       }
     })
@@ -82,34 +125,34 @@ describe('GeoEditableInput', () => {
   it('Should not render buttons if they are hidden', function () {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: false,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: false
       },
       stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
-        GeoDropdown
+        GeoDropdown,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton
       }
     })
 
-    expect(wrapper.find('.geo-editable-input__container__buttons__button--save').exists()).toBe(false)
-    expect(wrapper.find('.geo-editable-input__container__buttons__button--cancel').exists()).toBe(false)
+    expect(wrapper.find('geo-primary-compact-button').exists()).toBe(false)
+    expect(wrapper.find('geo-secondary-compact-button').exists()).toBe(false)
   })
 
   it('Should emit an event on click save button', function (done) {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: true,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: true
       },
       stubs: {
+        GeoDropdown,
         'font-awesome-icon': FontAwesomeIcon,
-        GeoDropdown
+        GeoCompactButton,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton
       }
     })
 
-    wrapper.find('.geo-editable-input__container__buttons__button--save').trigger('click')
+    wrapper.find(GeoPrimaryCompactButton).trigger('click')
     setTimeout(function () {
       try {
         expect(wrapper.emitted()['save']).toBeTruthy()
@@ -123,17 +166,18 @@ describe('GeoEditableInput', () => {
   it('Should emit an event on click cancel button', function (done) {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: true,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: true
       },
       stubs: {
+        GeoDropdown,
         'font-awesome-icon': FontAwesomeIcon,
-        GeoDropdown
+        GeoCompactButton,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton
       }
     })
 
-    wrapper.find('.geo-editable-input__container__buttons__button--cancel').trigger('click')
+    wrapper.find(GeoSecondaryCompactButton).trigger('click')
     setTimeout(function () {
       try {
         expect(wrapper.emitted()['cancel']).toBeTruthy()
@@ -144,40 +188,15 @@ describe('GeoEditableInput', () => {
     })
   })
 
-  it('Should show GeoActivityIndicator when it is loading', function (done) {
-    const wrapper = mount(GeoEditableInput, {
-      propsData: {
-        showButtons: true,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check'],
-        loading: true
-      },
-      stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
-        GeoActivityIndicator,
-        GeoDropdown
-      }
-    })
-    setTimeout(function () {
-      try {
-        expect(wrapper.contains(GeoActivityIndicator)).toBe(true)
-        done()
-      } catch (error) {
-        done(error)
-      }
-    })
-  })
-
   it('Should emit input event when added a value', function (done) {
     const wrapper = mount(GeoEditableInput, {
       propsData: {
-        showButtons: false,
-        cancelIcon: ['fas', 'times'],
-        saveIcon: ['fas', 'check']
+        showButtons: false
       },
       stubs: {
-        'font-awesome-icon': FontAwesomeIcon,
-        GeoDropdown
+        GeoDropdown,
+        GeoPrimaryCompactButton,
+        GeoSecondaryCompactButton
       }
     })